+++ src/resources/ts/components/organisms/Menubar.tsx
@@ -1,4 +1,4 @@
-import { Box, Stack, Text, VStack } from "@chakra-ui/react";
+import { Box, Text, VStack } from "@chakra-ui/react";
 import { memo } from "react";
 import { MenuList } from "../../MenuList";
 import { Menubox } from "../atomic/Menubox";
