+++ src/resources/ts/components/organisms/task/TaskAdd.tsx
@@ -5,7 +5,7 @@ import { ChangeEvent, memo, useEffect, useState } from "react";
 import { useRecoilState } from "recoil";
 import { iconManager } from "../../../icon";
 import { isChangedTaskAtom } from "../../../recoil/isChangedTaskAtom";
-import { loadingAtom } from "../../../recoil/isLoadingAtom";
+import { loadingAtom } from "../../../recoil/loadingAtom";
 import { newTaskAtom } from "../../../recoil/newTaskAtom";
 import { CancelButton } from "../../atomic/buttons/CancelButton";
 import { PrimaryButton } from "../../atomic/buttons/PrimaryButton";
