+++ src/resources/ts/components/organisms/task/TaskEditButtons.tsx
@@ -4,7 +4,7 @@ import { memo } from "react";
 import { useRecoilState } from "recoil";
 import { iconManager } from "../../../icon";
 import { isChangedTaskAtom } from "../../../recoil/isChangedTaskAtom";
-import { loadingAtom } from "../../../recoil/isLoadingAtom";
+import { loadingAtom } from "../../../recoil/loadingAtom";
 import { DeleteButton } from "../../atomic/buttons/DeleteButton";
 import { PrimaryButton } from "../../atomic/buttons/PrimaryButton";
 import { DeleteModal } from "../../molecules/DeleteModal";
